separate out HsTemplate and HsCast
